feature[test]: refactoring issue #507
