+++ mockk/common/src/test/kotlin/io/mockk/it/VerifyAtLeastAtMostExactlyTest.kt
@@ -1,6 +1,11 @@
 package io.mockk.it
 
-import io.mockk.*
+import io.mockk.Called
+import io.mockk.every
+import io.mockk.mockk
+import io.mockk.verify
+import io.mockk.verifyOrder
+import io.mockk.verifySequence
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertFailsWith
@@ -145,6 +150,29 @@ class VerifyAtLeastAtMostExactlyTest {
         }
     }
 
+    /**
+     * See issue #507
+     *
+     * A regression occurred in version 1.10.2 causing verify order to use
+     * eq() instead of any() matcher.
+     * This test exist to avoid this kind of regression in the future.
+     */
+    @Test
+    fun orderWithAny() {
+        val tracker = mockk<Tracker>(relaxUnitFun = true)
+        val player = Player(tracker)
+
+        player.goCrazy()
+
+        verifyOrder {
+            tracker.track(any(), "play", "param", any())
+            tracker.track(any(), "pause", "param", any())
+            tracker.track(any(), "play", "param", any())
+            tracker.track(any(), "pause", "param", any())
+            tracker.track(any(), "play", "param", any())
+        }
+    }
+
     @Test
     fun sequence() {
         doCalls()
@@ -188,4 +216,17 @@ class VerifyAtLeastAtMostExactlyTest {
         assertEquals(3, mock.op2(2, 1))
     }
 
+    interface Tracker {
+        fun track(song: String, action: String, param: String, moreParam: Map<String, String>)
+    }
+
+    class Player(private val tracker: Tracker) {
+        fun goCrazy() {
+            tracker.track("song 2", "play", "param", mapOf(Pair("key", "value")))
+            tracker.track("song 2", "pause", "param", mapOf(Pair("key", "value")))
+            tracker.track("song 2", "play", "param", mapOf(Pair("key", "value")))
+            tracker.track("song 2", "pause", "param", mapOf(Pair("key", "value")))
+            tracker.track("song 2", "play", "param", mapOf(Pair("key", "value")))
+        }
+    }
 }
